some clean up, added filling UI task list with tasks

+++ dog-walker-sim/Assets/Content/Scripts/Task.cs
@@ -10,7 +10,7 @@ public class Task
     [SerializeField] private Client client; public Client TaskClient { get { return client; } }
     [SerializeField] private Dog dog; public Dog TaskDog { get { return dog; } }
     [SerializeField] private GameObject address; public GameObject TaskAddress { get { return address; } }
-    [SerializeField] private float price; public float TaskPrice { get { return price; } }
+    [SerializeField] private float price; public float TaskPrice { get { return (int)price; } }
 
     public Task(Client client, Dog dog, float price)
     {
